src/phoenix_starter_kits/templates/llamaindex-qa/{{cookiecutter.project_slug}}/requirements.txt
arize-phoenix>=3.0.0
python-dotenv>=1.0.0
llama-index>=0.10.0
llama-index-llms-openai>=0.1.0
llama-index-embeddings-openai>=0.1.0